# Optimize notify of trigger listeners by only locking and allocating w…