--- a/src/Quartz.Benchmark/Directory.Build.props
+++ b/src/Quartz.Benchmark/Directory.Build.props
@@ -1,7 +1,0 @@
-<Project>
-  <Import Project="..\Directory.Build.props" />
-
-  <PropertyGroup>
-    <ProduceReferenceAssembly>false</ProduceReferenceAssembly>
-  </PropertyGroup>
-</Project>