… src/Quartz/Core/QuartzScheduler.cs