Optimize notify of trigger listeners by only locking and allocating when non-global trigger listeners have been added.
No longer define properties and methods for accessing or manipulating internal trigger listeners.

**.NET 4.8**

|                                                      Method | Branch |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------------------------------------ |-------:|-----------|----------:|----------:|-------:|----------:|
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | main   |  1322  ns | 2.62 ns   | 2.45 ns   | 0.1163 |     489 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | pr     |  936.7 ns | 3.30 ns   | 2.76 ns   | 0.0858 |     361 B |

**.NET 6.0**

|                                                      Method | Branch |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------------------------------------ |-------:|-----------|----------:|----------:|-------:|----------:|
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | main   | 902.0 ns  | 3.47 ns   | 3.24 ns   | 0.0668 |     280 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | pr     | 617.6 ns  | 1.12 ns   | 0.99 ns   | 0.0362 |     152 B |