@drieseng 2022-05-09
Marked as draft as I'm still working on adding more unit tests and improving documentation.