@drieseng 2022-05-14
@lahma This one is ready for review now.

Once this is merged, I see two remaining candidates for optimization for this specific scenario.
I'm just laying out further options to improve performance, not trying to convince you to take these changes.
The **ExecutingJobsManager** "proposal" is definitely not one to rush.

### Remove use of **Quartz.Core.Context**

*Context.CallerId* is an *AsyncLocal<Guid?>* that we assign a new **Guid** to in `JobRunShell.Run(CancellationToken cancellationToken = default)` and `QuartzSchedulerThread.Run()`.

The assigned value is used by **JobStoreSupport**. I can't immediately tell if the assign in `JobRunShell.Run(CancellationToken cancellationToken = default)` is actually necessary.

Removing it (branch 'experiment') gives us a nice win:

|                                                      Method | Branch |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|------------------------------------------------------------ |-------:| --------:|--------:|--------:|-------:|----------:|
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | 559fd40c29f3e63ca19659ee1e995c50440bec02 | 1566,0 ns | 1.14 ns | 1.01 ns | 0.2403 |   1,008 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | main   | 902.0 ns | 3.47 ns | 3.24 ns | 0.0668 |     280 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | pr     | 617.6 ns | 1.12 ns | 0.99 ns | 0.0362 |     152 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | experiment | 446.2 ns | 2.86 ns | 2.54 ns | 0.0114 |      48 B |

### Ship **ExecutingJobsManager** as an opt-in **IJobListener**

I'm not sure you would want to consider this as we'd have to remove the **NumJobsExecuted** and **CurrentlyExecutingJobs** properties from **QuartzScheduler**, and users would have to obtain this information directly from **ExecutingJobsManager**.

Doing this on top of the removal of / alternative for **Quartz.Core.Context** gets us to zero allocations and another nice performance boost:

|                                                      Method | Branch |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|------------------------------------------------------------ |-------:| --------:|--------:|--------:|-------:|----------:|
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | 559fd40c29f3e63ca19659ee1e995c50440bec02 | 1566,0 ns | 1.14 ns | 1.01 ns | 0.2403 |   1,008 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | main   | 902.0 ns | 3.47 ns | 3.24 ns | 0.0668 |     280 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | pr     | 617.6 ns | 1.12 ns | 0.99 ns | 0.0362 |     152 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | experiment | 301.0 ns | 2.77 ns | 2.45 ns | -      |         - |