@drieseng 2022-05-20
> Thanks for working out the PR. Generally I think reducing allocations for core services can be somewhat time wasted compared to the time building the actual scheduler can take.

Building the scheduler is a one-time cost. Paying extra for each job that is triggered is a cost that you pay time and time again.
In my project, we actually schedule millions of jobs a day. Personally, I don't consider my effort to be "time wasted".

Do you see a way to avoid the use of assigning of Context.CallerId in `JobRunShell.Run(CancellationToken cancellationToken = default)`?