@lahma 2022-05-20
> > Thanks for working out the PR. Generally I think reducing allocations for core services can be somewhat time wasted compared to the time building the actual scheduler can take.
> 
> Building the scheduler is a one-time cost. Paying extra for each job that is triggered is a cost that you pay time and time again. In my project, we actually schedule millions of jobs a day. Personally, I don't consider my effort to be "time wasted".

I think my comment came out a bit wrong. The thing was about trying to avoid allocation lists/dictionary instance members for core services which are instantiated only once per scheduler lifetime. Reducing recurring runtime allocations is of course awesome!

> Do you see a way to avoid the use of (assigning of) **Context.CallerId** in `JobRunShell.Run(CancellationToken cancellationToken = default)`?

Because it's used as locking handle in downstream ado job store it might be a bit tricky to get to flow downstream without changing all the APIs. I think `JobRunShell` and `QuartzSchedulerThread` would need to pass the value explicitly. Been quite the while I've been in contact with that async local so my memory isn't serving me the best here.